Move import commands to a separate package
