+++ dev/import-tool/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "@hcengineering/import-tool",
+  "version": "0.1.0",
+  "main": "lib/index.js",
+  "svelte": "src/index.ts",
+  "types": "types/index.d.ts",
+  "files": [
+    "lib/**/*",
+    "types/**/*",
+    "tsconfig.json"
+  ],
+  "author": "Anticrm Platform Contributors",
+  "scripts": {
+    "build": "compile",
+    "build:watch": "compile",
+    "_phase:bundle": "rushx bundle",
+    "bundle": "mkdir -p bundle && node esbuild.js",
+    "run-local": "cross-env SERVER_SECRET=secret MONGO_URL=mongodb://localhost:27017 FRONT_URL=http://localhost:8087  COLLABORATOR_URL=ws://localhost:3078 STORAGE_CONFIG=minio|minio?accessKey=minioadmin&secretKey=minioadmin node --nolazy -r ts-node/register ./src/__start.ts",
+    "run": "cross-env node -r ts-node/register --max-old-space-size=8000 ./src/__start.ts",
+    "format": "format src",
+    "test": "jest --passWithNoTests --silent",
+    "_phase:build": "compile transpile src",
+    "_phase:test": "jest --passWithNoTests --silent",
+    "_phase:format": "format src",
+    "_phase:validate": "compile validate"
+  },
+  "devDependencies": {
+    "cross-env": "~7.0.3",
+    "@hcengineering/platform-rig": "^0.6.0",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-n": "^15.4.0",
+    "eslint": "^8.54.0",
+    "ts-node": "^10.8.0",
+    "esbuild": "^0.20.0",
+    "@types/mime-types": "~2.1.1",
+    "@types/node": "~20.11.16",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint-config-standard-with-typescript": "^40.0.0",
+    "prettier": "^3.1.0",
+    "typescript": "^5.3.3",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "@types/jest": "^29.5.5"
+  },
+  "dependencies": {
+    "@hcengineering/attachment": "^0.6.14",
+    "@hcengineering/collaboration": "^0.6.0",
+    "@hcengineering/document": "^0.6.0",
+    "@hcengineering/text": "^0.6.0",
+    "@hcengineering/model-attachment": "^0.6.0",
+    "@hcengineering/model-core": "^0.6.0",
+    "@hcengineering/core": "^0.6.32",
+    "@hcengineering/platform": "^0.6.11",
+    "@hcengineering/server-tool": "^0.6.0",
+    "@hcengineering/server-client": "^0.6.0",
+    "commander": "^8.1.0",
+    "mime-types": "~2.1.34"
+  }
+}
